testNS/src/requirements.txt
appium-python-client
selenium
pytest